Fixed incorrect refactoring
