+++ src/main.rs
@@ -4,8 +4,6 @@ extern crate nom;
 mod parser;
 use parser::kb;
 
-mod knowledge_base;
-
 fn main() {
 
 }
